Move web3signer structs into their own file
